+++ ui/src/app/core/accessories/types/access-code/access-code.component.ts
@@ -1,0 +1,17 @@
+import { Component, Input } from '@angular/core'
+import { TranslatePipe } from '@ngx-translate/core'
+
+import { ServiceTypeX } from '@/app/core/accessories/accessories.interfaces'
+
+@Component({
+  selector: 'app-access-code',
+  templateUrl: './access-code.component.html',
+  standalone: true,
+  imports: [
+    TranslatePipe,
+  ],
+})
+export class AccessCodeComponent {
+  @Input() public service: ServiceTypeX
+  @Input() public readyForControl = false
+}